+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+## [5.1.0] - 2023-03-27
+### Added
+- new method for sign out `Client.signOut(ClientSignOutMode, Boolean)`. Old method is now deprecated.
+
+## [4.10.3] - 2023-03-27
+### Added
+- new method for sign out `Client.signOut(ClientSignOutMode, Boolean)`. Old method is now deprecated.
+
 ## [5.0.1] - 2023-03-14
 ### Added
 - new context passing mechanism based on androidx.startup library